Score items'stitle aligned center, paddid behind residetnt added, global component added and animations modified
